Prog3 done

+++ sem4/oop/prog1/src/com/company/Main.java
@@ -11,7 +11,7 @@ public class Main {
 
         Scanner s1 = new Scanner(System.in);
         Complex c1 = new Complex(2, 4);    //Using the constructor to initialize
-        Complex c2 = new Complex(s1.nextInt(), s1.nextInt());
+        Complex c2 = new Complex(s1.nextDouble(), s1.nextDouble());
         Complex c3 = new Complex();
         System.out.println("Initially :\nc1\t\t\tc2\t\t\tc3");
         c1.display();c2.display();c3.display();
